+++ R/nrmse.R
@@ -2,7 +2,7 @@
 # Part of the hydroGOF R package, https://github.com/hzambran/hydroGOF
 #                                 https://cran.r-project.org/package=hydroGOF
 #                                 http://www.rforge.net/hydroGOF/ ;
-# Copyright 2008-2023 Mauricio Zambrano-Bigiarini
+# Copyright 2008-2025 Mauricio Zambrano-Bigiarini
 # Distributed under GPL 2 or later
 
 ################################################################################
@@ -14,33 +14,36 @@
 # Updates: 06-Sep-2009                                                         #
 #          03-Jul-2017                                                         #
 #          16-Jan-2023                                                         #
+#          21-Dec-2025                                                         #
 ################################################################################
 # 'obs'   : numeric 'data.frame', 'matrix' or 'vector' with observed values
 # 'sim'   : numeric 'data.frame', 'matrix' or 'vector' with simulated values
 
 # 'norm'  : character, indicating the value to be used to normalise the RMS. Valid values are:
-#           -) 'sdobs' : standard deviation of observations.
+#           -) 'sd'    : standard deviation of observations.
 #           -) 'maxmin': difference between maximum and minimum observed values
+#           -) 'mean'  : arithmetic mean of observed values
+#           -) 'IQR'   : interquantile range of observed values, computed using the \code{stats::IQR} function with \code{type=7}.
 
 # 'Result': Normalized Root Mean Square Error between 'sim' and 'obs', 
 #           when multiplied by 100 its units is %
 
 nrmse <-function(sim, obs, ...) UseMethod("nrmse")
  
-nrmse.default <- function(sim, obs, na.rm=TRUE, norm="sd", fun=NULL, ...,
+nrmse.default <- function(sim, obs, na.rm=TRUE, norm=c("sd", "maxmin", "mean", "IQR"), fun=NULL, ...,
                           epsilon.type=c("none", "Pushpalatha2012", "otherFactor", "otherValue"), 
                           epsilon.value=NA) {
 
-  # Checking that the user provied a valid argument for 'norm'       
-  if (is.na(match(norm, c("sd", "maxmin") ) ) ) 
-       stop("Invalid argument: 'norm' must be in c('sd', 'maxmin')")
-
   if ( is.na(match(class(sim), c("integer", "numeric", "ts", "zoo", "xts"))) |
        is.na(match(class(obs), c("integer", "numeric", "ts", "zoo", "xts")))
-     ) stop("Invalid argument type: 'sim' & 'obs' have to be of class: c('integer', 'numeric', 'ts', 'zoo', 'xts')")      
+     ) stop("Invalid argument type: 'sim' & 'obs' have to be of class: c('integer', 'numeric', 'ts', 'zoo', 'xts')")     
 
-  epsilon.type <- match.arg(epsilon.type)  
+  # Checking the 'norm' argument
+  norm <- match.arg(norm)   
 
+  # Checking the 'epsilon.type' argument
+  epsilon.type <- match.arg(epsilon.type) 
+ 
   # index of those elements that are present both in 'sim' and 'obs' (NON- NA values)
   vi <- valindex(sim, obs)
    
@@ -57,27 +60,30 @@ nrmse.default <- function(sim, obs, na.rm=TRUE, norm="sd", fun=NULL, ...,
       sim  <- new[["sim"]]
       obs  <- new[["obs"]]
     } # IF end     
-       
-    if (norm=="sd") {
-      cte <- sd(obs, na.rm=na.rm)
-    } else if (norm=="maxmin") {
-        cte <- ( max(obs, na.rm= na.rm) - min(obs, na.rm =na.rm) )
+
+    cte = switch(  
+            norm,  
+            "sd"= sd(obs, na.rm=na.rm), 
+            "maxmin"= ( max(obs, na.rm= na.rm) - min(obs, na.rm =na.rm) ),  
+            "mean"= mean(obs, na.rm=na.rm),  
+            "IQR"= stats::IQR(obs, na.rm= na.rm, type=7),
+          )  
+
+    rmse <- rmse(sim, obs, na.rm) 
+     
+    if (cte != 0) {     
+       nrmse <- rmse / cte     
+    } else {
+        nrmse <- NA
+        warning("'obs' is constant -> it is not possible to compute 'nrmse' !")  
       } # ELSE end
 
-     rmse <- rmse(sim, obs, na.rm) 
+  } else {
+      nrmse <- NA
+      warning("There are no pairs of 'sim' and 'obs' without missing values !")
+    } # ELSE end
      
-     if (max(obs, na.rm= na.rm) - min(obs, na.rm= na.rm) != 0) {     
-       nrmse <- rmse / cte     
-     } else {
-         nrmse <- NA
-         warning("'obs' is constant -> it is not possible to compute 'nrmse' !")  
-       } # ELSE end
-   } else {
-         nrmse <- NA
-         warning("There are no pairs of 'sim' and 'obs' without missing values !")
-     } # ELSE end
-     
-   return( round( 100*nrmse, 1) )
+  return( round( 100*nrmse, 1) )
      
 } # 'nrmse.default' end
   
@@ -89,8 +95,9 @@ nrmse.default <- function(sim, obs, na.rm=TRUE, norm="sd", fun=NULL, ...,
 ################################################################################
 # Started: 15-Dic-2008                                                         #
 # Updates: 06-Sep-2009 ; 05-Nov-2012                                           #
+#          21-Dec-2025                                                         #
 ################################################################################
-nrmse.matrix <- function(sim, obs, na.rm=TRUE, norm="sd", fun=NULL, ...,
+nrmse.matrix <- function(sim, obs, na.rm=TRUE, norm=c("sd", "maxmin", "mean", "IQR"), fun=NULL, ...,
                          epsilon.type=c("none", "Pushpalatha2012", "otherFactor", "otherValue"), 
                          epsilon.value=NA) {
 
@@ -99,10 +106,6 @@ nrmse.matrix <- function(sim, obs, na.rm=TRUE, norm="sd", fun=NULL, ...,
     stop( paste("Invalid argument: dim(sim) != dim(obs) ( [", 
           paste(dim(sim), collapse=" "), "] != [", 
           paste(dim(obs), collapse=" "), "] )", sep="") )
-          
-  # Checking that the user provied a valid argument for 'norm'       
-  if (is.na(match(norm, c("sd", "maxmin") ) ) ) 
-     stop("Invalid argument: 'norm' must be in c('sd', 'maxmin')")
 
   nrmse <- rep(NA, ncol(obs))       
           
@@ -127,8 +130,9 @@ nrmse.matrix <- function(sim, obs, na.rm=TRUE, norm="sd", fun=NULL, ...,
 ################################################################################
 # Started: 15-Dic-2008                                                         #
 # Updates: 06-Sep-2009                                                         #
+#          21-Dec-2025                                                         #
 ################################################################################
-nrmse.data.frame <- function(sim, obs, na.rm=TRUE, norm="sd", fun=NULL, ...,
+nrmse.data.frame <- function(sim, obs, na.rm=TRUE, norm=c("sd", "maxmin", "mean", "IQR"), fun=NULL, ...,
                              epsilon.type=c("none", "Pushpalatha2012", "otherFactor", "otherValue"), 
                              epsilon.value=NA) {
 
@@ -145,9 +149,9 @@ nrmse.data.frame <- function(sim, obs, na.rm=TRUE, norm="sd", fun=NULL, ...,
 # Author: Mauricio Zambrano-Bigiarini                                          #
 ################################################################################
 # Started: 22-Mar-2013                                                         #
-# Updates:                                                                     #
+# Updates: 21-Dec-2025                                                         #
 ################################################################################
-nrmse.zoo <- function(sim, obs, na.rm=TRUE, norm="sd", fun=NULL, ...,
+nrmse.zoo <- function(sim, obs, na.rm=TRUE, norm=c("sd", "maxmin", "mean", "IQR"), fun=NULL, ...,
                       epsilon.type=c("none", "Pushpalatha2012", "otherFactor", "otherValue"), 
                       epsilon.value=NA){