+++ NEWS.md
@@ -5,7 +5,10 @@ NEWS/ChangeLog for hydroGOF
 ## Bug fixes
         o 'APFB' : Its value is now correctly computed, previously it was not Thanks to Rodrigo Mariano and Héctor Garcés
 
-
+## Enhancements
+        o 'plot2': now it shows and adaptive time axis for sub-daily zoo objects.
+        o 'ggof' : now it shows and adaptive time axis for sub-daily zoo objects.
+        o 'gof'  : now it does not compute sKGE, APFB and HFB when sim and obs are annual zoo objects.
 
 # Changes in version 0.6-0  08-May-2024